Move media report admins to a separate file
